balancer: implement actions using new base action
