+++ packages/balancer-fee-collector/contracts/interfaces/IBalancerBoostedPool.sol
