+++ packages/balancer-fee-collector/contracts/interfaces/IBalancerPool.sol
