+++ packages/balancer-fee-collector/package.json
@@ -22,11 +22,11 @@
     "prepare": "yarn build"
   },
   "dependencies": {
-    "@mimic-fi/v2-smart-vaults-base": "0.0.8",
+    "@mimic-fi/v2-smart-vaults-base": "0.0.9",
     "@openzeppelin/contracts": "4.7.0"
   },
   "devDependencies": {
-    "@mimic-fi/v2-helpers": "0.1.1",
+    "@mimic-fi/v2-helpers": "0.1.3",
     "@nomiclabs/hardhat-ethers": "^2.2.1",
     "@nomiclabs/hardhat-waffle": "^2.0.3",
     "@types/chai": "^4.2.18",
